Check unique windows by their whole title
The previous revision of `main.cpp` only checked the uniqueness of a window based on the address or name of the remote host. This commit changes the implementation to compare the whole title for better accuracy.
---
There were also very slight modifications to `Console::update()` in `uicomp.cpp` to initialize an ImVec4 using brace initialization `{ ... }` rather than the more verbose constructor `ImVec4(...)`.