Remove newlines in ClientWindow titles
A newline can be part of a Bluetooth device name, which is shown in the title of a Bluetooth connection window. If a name with newline(s) present is shown, anything after the first will get clipped.
---
This commit changes the `UIHelpers::makeClientWindowTitle` function to replace newlines with spaces before returning the formatted title string.